+++ root/root_test.go
@@ -5,4 +5,11 @@ import "testing"
 func TestRoot(t *testing.T) {
 	// TODO: NewCommand
 	// TODO: NewCommandMap
+	// TODO: MainLoop
+	// TODO: ParamsMap
+	// TODO: Authenticate
+	// TODO: validateCommandParams
+	// TODO: getCommandParams
+	// TODO: displayHelp
+	// TODO: searchCommand
 }